Added react-router, Login.js
